More type shenanigans

+++ src/osk.hpp
@@ -3,15 +3,15 @@
 #include "../FBInk/fbink.h"
 #include "_kblayout.hpp"
 
-void osk_setup(int width, int height) {
-    int blockw = width / OSK_W;
-    int blockh = height / OSK_H;
-    int bpp = 1;
-    float radius = 10;
-    float spacing = 3;
-    for (int i = 0; i < OSK_NKEYS; i++) {
-        osk_keys[i].rrect.width = osk_keys[i].w * blockw;
-        osk_keys[i].rrect.height = osk_keys[i].h * blockh;
+void osk_setup(unsigned int width, unsigned int height) {
+    unsigned int blockw = width / OSK_W;
+    unsigned int blockh = height / OSK_H;
+    uint8_t bpp = 1u;
+    float radius = 10.f;
+    float spacing = 3.f;
+    for (unsigned int i = 0; i < OSK_NKEYS; i++) {
+        osk_keys[i].rrect.width = static_cast<unsigned int>(osk_keys[i].w * static_cast<float>(blockw));
+        osk_keys[i].rrect.height = static_cast<unsigned int>(osk_keys[i].h * static_cast<float>(blockh));
         osk_keys[i].rrect.bpp = bpp;
         osk_keys[i].rrect.radius = radius;
         osk_keys[i].rrect.spacing = spacing;
@@ -20,18 +20,18 @@ void osk_setup(int width, int height) {
     }
 }
 
-void osk_render(int fd, FBInkConfig * config, int osk_y, int width, int height) {
-    int blockw = width / OSK_W;
-    int blockh = height / OSK_H;
+void osk_render(int fd, FBInkConfig * config, unsigned int osk_y, unsigned int width, unsigned int height) {
+    unsigned int blockw = width / OSK_W;
+    unsigned int blockh = height / OSK_H;
     short cfg_row = config->row;
     short cfg_col = config->col;
     config->row = 0;
     config->col = 0;
     // Batch it
     config->no_refresh = true;
-    for (int i = 0; i < OSK_NKEYS; i++) {
-        int x = osk_keys[i].x * blockw;
-        int y = osk_keys[i].y * blockh;
+    for (unsigned int i = 0; i < OSK_NKEYS; i++) {
+        unsigned int x = static_cast<unsigned int>(osk_keys[i].x * static_cast<float>(blockw));
+        unsigned int y = static_cast<unsigned int>(osk_keys[i].y * static_cast<float>(blockh));
         // printf("%d %d %d %d %d\n", i, x, y, osk_keys[i].rrect.width, osk_keys[i].rrect.height);
         fbink_print_raw_data(
                 fd,
@@ -39,8 +39,8 @@ void osk_render(int fd, FBInkConfig * config, int osk_y, int width, int height) 
                 osk_keys[i].rrect.width,
                 osk_keys[i].rrect.height,
                 osk_keys[i].rrect.width * osk_keys[i].rrect.height * osk_keys[i].rrect.bpp,
-                x,
-                osk_y + y,
+                static_cast<short int>(x),
+                static_cast<short int>(osk_y + y),
                 config
                 );
     }
@@ -52,14 +52,14 @@ void osk_render(int fd, FBInkConfig * config, int osk_y, int width, int height) 
     config->col = cfg_col;
 }
 
-const kbkey * osk_press(int width, int height, int x, int y) {
-    int blockw = width / OSK_W;
-    int blockh = height / OSK_H;
-    for (int i = 0; i < OSK_NKEYS; i++) {
-        int kx = osk_keys[i].x * blockw;
-        int ky = osk_keys[i].y * blockh;
-        int kw = osk_keys[i].rrect.width;
-        int kh = osk_keys[i].rrect.height;
+const kbkey * osk_press(unsigned int width, unsigned int height, unsigned int x, unsigned int y) {
+    unsigned int blockw = width / OSK_W;
+    unsigned int blockh = height / OSK_H;
+    for (unsigned int i = 0; i < OSK_NKEYS; i++) {
+        unsigned int kx = static_cast<unsigned int>(osk_keys[i].x * static_cast<float>(blockw));
+        unsigned int ky = static_cast<unsigned int>(osk_keys[i].y * static_cast<float>(blockh));
+        unsigned int kw = osk_keys[i].rrect.width;
+        unsigned int kh = osk_keys[i].rrect.height;
         if (x >= kx && x < kx + kw && y >= ky && y < ky + kh) {
             return &osk_keys[i];
         }
